# binary/reader: Skip fixed witdth collections faster